Extract exportTest to separate class
